add case insentive checks
There are more than one way in which case insensitive checks can be handled.

I decided to make the repository case insensitive when creating it.

Pros:
* it is fast to check
* there is no need to normalize after the fact (you might have to combine separate entries eg)
* the repository represents the actual base against which we will compare

Cons:
* the repository does not represent the code as it is read by the parser
* old repositories are not compatible with the new ones

An alternative approach would be to normalize the repository before checking. This has the
already mentioned drawbacks.

A third alternative would be to handle the case insensitive check in the actual node visitor.
This would be extremely expensive and as this is done a lot of times during a normal
check of a code base, this is not what I would suggest albeit it would probably be the
'cleanest' solution as it actually only touches the check itself and nothing else.

Fixes #12